requirement.txt
absl-py==1.4.0
aiohttp==3.8.4
aiosignal==1.3.1
async-timeout==4.0.2
attrs==22.2.0
blinker==1.6.2
cachetools==5.3.0
# certifi @ file:///croot/certifi_1671487769961/work/certifi
charset-normalizer==3.1.0
click==8.1.6
cupy-cuda11x==12.3.0
cycler==0.11.0
einops==0.6.1
et-xmlfile==1.1.0
faiss-cpu==1.7.2
faiss-gpu==1.7.2
fastrlock==0.8.2
filelock==3.10.6
fire==0.5.0
Flask==2.3.2
fonttools==4.39.2
frozenlist==1.3.3
fsspec==2023.3.0
google-auth==2.16.3
google-auth-oauthlib==0.4.6
googledrivedownloader==0.4
grpcio==1.51.3
h5py==3.8.0
huggingface-hub==0.0.12
idna==3.4
importlib-metadata==6.1.0
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.2.0
kiwisolver==1.4.4
lightning-utilities==0.3.0
loguru==0.7.2
Markdown==3.4.3
MarkupSafe==2.1.2
matplotlib==3.5.3
multidict==6.0.4
natsort==8.3.1
numpy==1.23.4
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
oauthlib==3.2.2
opencv-python==4.7.0.72
openpyxl==3.1.2
packaging==23.0
pandas==1.5.1
Pillow==9.2.0
prettytable==3.5.0
protobuf==3.20.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycocotools==2.0.7
pyparsing==3.0.9
python-dateutil==2.8.2
pytorch-gpu==0.0.1
pytorch-lightning==1.8.3
pytorch-metric-learning==1.6.3
pytz==2023.2
PyYAML==6.0
regex==2023.6.3
requests==2.28.2
requests-oauthlib==1.3.1
rsa==4.9
sacremoses==0.0.53
scikit-learn==1.2.1
scipy==1.9.3
six==1.16.0
# sklearn==0.0.post1
tabulate==0.9.0
tensorboard==2.11.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorboardX==2.6
termcolor==2.2.0
thop==0.1.1.post2209072238
threadpoolctl==3.1.0
timm==0.6.12
tokenizers==0.10.3
torch==1.13.1
torchaudio==0.13.1
torchmetrics==0.11.1
torchscan==0.1.1
torchvision==0.14.1
tqdm
transformers==4.8.2
typing_extensions==4.5.0
urllib3==1.26.15
wcwidth==0.2.6
Werkzeug==2.3.6
yacs==0.1.8
yarl==1.8.2
zipp==3.15.0
